+++ probirka/_ext/aiohttp.py
@@ -1,6 +1,9 @@
+from __future__ import annotations
+
 import json
 
-from typing import Any, Callable, Coroutine, List, Optional, Union
+from collections.abc import Callable, Coroutine, Sequence
+from typing import Any
 
 from aiohttp import web
 
@@ -9,8 +12,8 @@ from probirka._probirka import Probirka
 
 def make_aiohttp_endpoint(
     probirka: Probirka,
-    timeout: Optional[int] = None,
-    with_groups: Union[str, List[str]] = '',
+    timeout: int | None = None,
+    with_groups: str | Sequence[str] = '',
     skip_required: bool = False,
     return_results: bool = True,
     success_code: int = 200,
@@ -21,8 +24,8 @@ def make_aiohttp_endpoint(
 
     Args:
         probirka (Probirka): The Probirka instance to run.
-        timeout (Optional[int]): The timeout for the Probirka run.
-        with_groups (Union[str, List[str]]): Groups to include in the Probirka run.
+        timeout (int | None): The timeout for the Probirka run.
+        with_groups (str | Sequence[str]): Groups to include in the Probirka run.
         skip_required (bool): Whether to skip required checks.
         return_results (bool): Whether to return the results in the response.
         success_code (int): The HTTP status code for a successful response.